Do 05-flex-modal
Spent considerably more time compared to previous exercies. To have the blue icon to the left be where it is, need to be in it's own container then the rest of the content in another one. The other container contains everything else. Had to set flex-shrink to 0 for the container containing the blue icon too.